modified obj_cl to opt only on xi, a and release of obj_geom_eta
modifications have been added as a low robustness was found. With the previous version of obj_cl the eta parameter (not optimized) was clearly not robust. Therefore, after optimizing for xi, a through the new version of obj_cl the code then optimizes for eta through obj_geom_eta, basically the same function as obj_geom but with fixed xi and a parameters. This has been proven to add robustness to the final output airfoils. Note that initial guesses MUST be reasonable, for 2 reasons: 1) fminsearch is very sensible to them in this case 2) cl is calculated with the approximation of small xi/a, which is clearly true for the NACA23012; therefore one should use initial guesses that respect this.